refactor(worker): move WorkerInstance into a separate module
affects: @zetapush/worker
