+++ packages/worker/src/client/worker.js
@@ -1,62 +1,7 @@
 import { Authentication, Client, uuid } from '@zetapush/core';
 import { Queue as Worker } from '@zetapush/platform';
 
-import { timeoutify } from '../utils/async.js';
-
-class WorkerInstance {
-  constructor({ timeout, worker }) {
-    /**
-     * @access private
-     * @type {number}
-     */
-    this.timeout = timeout;
-    /**
-     * @access private
-     * @type {Object}
-     */
-    this.worker = worker;
-  }
-  async dispatch({ data: { request, taskId } }) {
-    const { data, requestId, owner } = request;
-    const { name, namespace, parameters } = data;
-    console.log('WorkerInstance::dispatch', {
-      name,
-      namespace,
-      parameters,
-      requestId,
-      taskId,
-    });
-    try {
-      // Wrap request context
-      const context = {
-        owner,
-      };
-      const result = await timeoutify(
-        () => this.worker[namespace][name](parameters, context),
-        this.timeout,
-      );
-      console.log('WorkerInstance::result', result);
-      return {
-        result,
-        taskId,
-        requestId,
-        success: true,
-      };
-    } catch (error) {
-      console.error('WorkerInstance::error', error);
-      return {
-        result: error,
-        taskId,
-        requestId,
-        success: false,
-      };
-    }
-  }
-  setWorker(worker) {
-    console.log('WorkerInstance::setWorker', worker);
-    this.worker = worker;
-  }
-}
+import { WorkerInstance } from '../utils/worker-instance.js';
 
 export class WorkerClient extends Client {
   constructor({
