+++ packages/worker/src/utils/worker-instance.js
@@ -1,0 +1,56 @@
+import { timeoutify } from './async.js';
+
+export class WorkerInstance {
+  constructor({ timeout, worker }) {
+    /**
+     * @access private
+     * @type {number}
+     */
+    this.timeout = timeout;
+    /**
+     * @access private
+     * @type {Object}
+     */
+    this.worker = worker;
+  }
+  async dispatch({ data: { request, taskId } }) {
+    const { data, requestId, owner } = request;
+    const { name, namespace, parameters } = data;
+    console.log('WorkerInstance::dispatch', {
+      name,
+      namespace,
+      parameters,
+      requestId,
+      taskId,
+    });
+    try {
+      // Wrap request context
+      const context = {
+        owner,
+      };
+      const result = await timeoutify(
+        () => this.worker[namespace][name](parameters, context),
+        this.timeout,
+      );
+      console.log('WorkerInstance::result', result);
+      return {
+        result,
+        taskId,
+        requestId,
+        success: true,
+      };
+    } catch (error) {
+      console.error('WorkerInstance::error', error);
+      return {
+        result: error,
+        taskId,
+        requestId,
+        success: false,
+      };
+    }
+  }
+  setWorker(worker) {
+    console.log('WorkerInstance::setWorker', worker);
+    this.worker = worker;
+  }
+}
